Force the DOM parser to continue with errors (#808)
* Force the DOM parser to continue with compilation errors
