+++ org.eclipse.jdt.core.javac/src/org/eclipse/jdt/internal/javac/JavacCompiler.java
@@ -19,7 +19,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
-import java.util.Queue;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -46,12 +45,9 @@ import org.eclipse.jdt.internal.core.JavaProject;
 import org.eclipse.jdt.internal.core.builder.SourceFile;
 
 import com.sun.tools.javac.api.MultiTaskListener;
-import com.sun.tools.javac.comp.*;
 import com.sun.tools.javac.comp.CompileStates.CompileState;
 import com.sun.tools.javac.main.JavaCompiler;
-import com.sun.tools.javac.tree.JCTree.JCClassDecl;
 import com.sun.tools.javac.util.Context;
-import com.sun.tools.javac.util.Pair;
 
 public class JavacCompiler extends Compiler {
 	JavacConfig compilerConfig;
@@ -117,46 +113,7 @@ public class JavacCompiler extends Compiler {
 			// Configure Javac to generate the class files in a mapped temporary location
 			var outputDir = JavacClassFile.getMappedTempOutput(outputSourceSet.getKey()).toFile();
 			JavacUtils.configureJavacContext(javacContext, this.compilerConfig, javaProject, outputDir, true);
-			JavaCompiler javac = new JavaCompiler(javacContext) {
-				boolean isInGeneration = false;
-
-				@Override
-				protected boolean shouldStop(CompileState cs) {
-					// Never stop
-					return false;
-				}
-
-				@Override
-				public void generate(Queue<Pair<Env<AttrContext>, JCClassDecl>> queue, Queue<JavaFileObject> results) {
-					try {
-						this.isInGeneration = true;
-						super.generate(queue, results);
-					} catch (Throwable ex) {
-						// TODO error handling
-					} finally {
-						this.isInGeneration = false;
-					}
-				}
-
-				@Override
-				protected void desugar(Env<AttrContext> env, Queue<Pair<Env<AttrContext>, JCClassDecl>> results) {
-					try {
-						super.desugar(env, results);
-					} catch (Throwable ex) {
-						// TODO error handling
-					}
-				}
-
-				@Override
-				public int errorCount() {
-					// See JavaCompiler.genCode(Env<AttrContext> env, JCClassDecl cdef),
-					// it stops writeClass if errorCount is not zero.
-					// Force it to return 0 if we are in generation phase, and keeping
-					// generating class files for those files without errors.
-					return this.isInGeneration ? 0 : super.errorCount();
-				}
-			};
-			javacContext.put(JavaCompiler.compilerKey, javac);
+			JavaCompiler javac = TolerantJavaCompiler.configureCompilerInstance(javacContext);
 			javac.shouldStopPolicyIfError = CompileState.GENERATE;
 			try {
 				javac.compile(com.sun.tools.javac.util.List.from(outputSourceSet.getValue().stream()
